BB implementation of MIMO sysid
Its finally here. Main routine is atomic_LTI_BB, which calls BB_operator. The linear operators are A and A' in mimo_A and mimo_At. exTrace_BB is a modification of Rajiv's exTrace for the purposes of testing bag and bash. exMIMO_basic_BB is the script to run the examples, once again a slight modification of exMIMO_basic.

The BB code works to generate some result, but I'm not sure how to properly assess quality. I don't know what is happening at and around line 52 of exTrace_BB, which is where the benchmarks are evaluated. Hopefully Rajiv can help.