Fixed incorrect calculation of time (impulse) response
Also indexing of groups were incorrect. This is now the original order, which is less amenable towards randomization. Will need to switch over orderings later, in which the pole is  in the last dimension instead of the first.

There is therefore no reason to keep mimo_A and mimo_A2 both around. I have eliminated this redundancy, and changed the name to mimo_A everywhere.