Bugfixing on reweighted heuristic
Check out weight_compare in atomic_LTI_BB. This is the weighted atomic norm before and after reweighting. If the error goes down or stays the same on the reweighting  iter ation, then the current atomic norm = tau. If the error goes up, then the current atomic norm is less than tau. Strangely, this is not reflected in the BB code, what I see is that BB is on the boundary (run_log.warm_start.bash_manager.update_ext.exterior = 1), but the penalty as computed in atomic_LTI_BB does not agree.

Really have no idea what is going on. Will soon try spectra for groups and try and figure out why this is occuring. Makes no sense.