Tracked down bug in reweighted heuristic
This is actually more of an existential bug in optimization, and may be caused by a bug in gradient computation or setting weights.

What is observed here is that a set of atoms in a group have all coordinates changing signs. If all atoms have the same sign in a single coordinate, then the L infty norm (max absolute value) by taking the sum of the weights. As soon as the streak is broken, then some amount of coefficient mass is wasted (?).

The full atomic norm budget is therefore inefficiently used? Not really sure how to fix this. The gradient is adding an atom opposite to the streak.