Tau slippage
With some values of weighting functions, BB gets stuck. Specifically, BB (greedily) chooses a bad atom (or rather a good atom considering a history of  bad atoms), and then coefficient mass (atomic norm  budget) is wasted compensating for this new atom. This is apparent in the reweighting example ring_broken_test.mat.

This shows up in reweighting the ring examples, when reweighting error increases. The inefficient atomic norm allocation produces a point on the interior, and then a skewed reweighted face is produced around the interior point. This caused the spike in error due to bad weights.

One approach is to regenerate a set of new atoms when BB  gets stuck (cheat to win), see atom_repick.m for an example. What I did was to allow tau to slip. This isn't very kosher, but does the job for now. Will implement regeneration soon.